Make compliance dashboard a separate component
